+++ rustport/cppshim/include/bindings.hpp
@@ -4,22 +4,9 @@
 #include <type_traits>
 #include <utility>
 
-namespace stlab {
-inline namespace v1 {
-namespace detail {
+namespace rust {
 
-enum class executor_priority { high, medium, low };
-
-inline auto bridge_priority(executor_priority p) -> Priority {
-    switch (p) {
-        case executor_priority::high: return Priority::High;
-        case executor_priority::medium: return Priority::Default;
-        case executor_priority::low: return Priority::Low;
-        default: return Priority::Default;
-    }
-}
-
-/// @brief Asynchronously invoke f on the default_executor.
+/// @brief Asynchronously invoke f on the rust-backed executor.
 /// @tparam F function object type
 /// @param f a function object.
 /// @return the result of calling `execute`.
@@ -34,45 +21,20 @@ auto enqueue(F f) {
     });
 }
 
-/// @brief Asynchronously invoke `f` on the default_executor with priority `p`.
+/// @brief Asynchronously invoke `f` on the rust-backed executor with priority `p`.
 /// @tparam F function object type
 /// @param f a function object.
 /// @param priority 
 /// @return the value returned by `execute_priority`.
 template <class F, class = std::enable_if_t<std::is_invocable_r< void, F >::value> >
-auto enqueue_priority(F f, executor_priority p) {
+auto enqueue_priority(F f, Priority p) {
     return execute_priority(new F(std::move(f)), [](void* f_) {
         auto f = static_cast<F*>(f_);
         try {
             (*f)();
         } catch (...) {}
         delete f;
-    }, bridge_priority(p));
+    }, p);
 }
 
-/// @brief A thin invokable wrapper around `enqueue_priority`.
-/// @tparam Priority the priority at which all given function objects will be enqueued.
-template <executor_priority Priority>
-struct executor_type {
-    using result_type = void;
-
-    /// @brief Enqueues the given task on the default_executor with this object's Priority value.
-    /// @tparam F function object type
-    /// @param f function object
-    template <class F>
-    void operator()(F&& f) const {
-        enqueue_priority(std::forward<F>(f), Priority);
-    }
-};
-
-} // namespace detail
-
-/// @brief An executor for low priority tasks, enqueued with the call operator.
-constexpr auto low_executor = detail::executor_type<detail::executor_priority::low>{};
-/// @brief An executor for standard priority tasks, enqueued with the call operator.
-constexpr auto default_executor = detail::executor_type<detail::executor_priority::medium>{};
-/// @brief An executor for high priority tasks, enqueued with the call operator.
-constexpr auto high_executor = detail::executor_type<detail::executor_priority::high>{};
-
-} // inline namespace v1
-} // namespace stlab
+}
